Steal focus when interacting with the slider
A div can't get focus, at least in Chrome, but it makes sense for e.g. an input box to lose focus when a user starts to interact with the slider.

This is necessary for my app, where the slider and the text box tightly interact, and is also consistent with jQueryUI Slider.